Include LUA 5.4.3 and compile it as C++
[why]
The LUA distributions that come bundled with the
operating system(s) are all built as pure C libraries.
In this mode, LUA uses longjmp() to handle errors.
Unfortunately, this conflicts with the proper stack
unwinding (especially running of destructors) in C++
code, which is especially relevant for our timeout
debug hook function, but also for future C++ functions
that should be called from LUA.

If LUA is built with a C++ compiler, it automatically
uses C++ exceptions instead of longjmp(), which solves
this problem. Apart from that, we get a newer version
of LUA than typically available in our OS distros.

[how]
Include the LUA sources from
https://www.lua.org/download.html,
rename all .c files to .cc, and build a simple static
library from them with Meson. This also drops the
build-time dependency on lua5.3.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>