execute_step(): Make timeout more reliable
[why]
A timeout generated a normal LUA error that could
be intercepted by a script via pcall(). This allowed
users to write infinite loops that could not be
terminated.

[how]
When throwing a timeout error, also install a new
debug hook that is called whenever LUA execution is
resumed. This effectively leads to the throwing of
errors until all pcall()s have exited and execution
is finished.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>