Sequence: Add automatic indentation
[why]
Indentation is helpful for humans to understand the
nesting of block structures such as IF, WHILE, TRY.
It is also helpful to verify the correctness of a
Sequence.

[how]
On every modification of the sequence steps, indent()
is called and assigns an indentation level to all steps.
It also detects any inconsistencies; if there are any,
they can be retrieved as an error message with
get_indentation_error().

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>