deb: Use project default Meson settings
[why]
It can be quite confusing that we have two places to adjust the install
paths and so on. Usually people might think that packages will be
created like the meson.build default settings.

I see no reason why this shall be different?

Especially commit 0624c0b2b7 changed the prefix, but this branch
re-introduces the old prefix again, independently.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>