deb: Fix package path move
[why]
When the package moves from /export/doocs to /usr we also need to change
the packaging paths accordingly.

[how]
Also skip headers that were installed directly into /usr/install, as
that is not nice and should be avoided. If we ever produce such headers
we will get a warning (and them missing from the package).

Also fix the libs directory, because debian wants to put it somewhere we
do not ;-)

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>