deb: Remove created files on dh_clean
[why]
We create a file by an external script, and so we must clean up
afterwards ourselves.

[how]
It's a bit tricky because we have the version number in the lib install
file, and a simple glob also matches the dev-lib's install file.

So we use a number of '?' to make sure we do not match "-dev".

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>