Improved the syntax checker and interpreter on nested steps.
It works now properly as you can see from the test, especially
test_execute_sequence.cc. However the execute_sequencer.cc still has
function to find the required types which can be solved more fluently by
using on generic function with lambdas - next merge request ;)

The syntax checker and the interpreter are now improved with recursive
function calls. About the interpreter I am not to 100% happy with
the solution (see execute_sequencer.cc) as it is a little bit complex
to handle all of the different step types, espcially for the try block
that executes a recursive block where we need to transfer the actual
'try'-variables ... I am sure that there are better solutions but
I would suppose it needs a complete new approach.