tests: Use dependency instead of relative path for include
[why]
Normally tests should not rely on anything that the later library user
can not (so that the tests can work as examples).

Furthermore the include directories are already set up, so relative
paths to the header files makes not really sense.

[how]
Rely on meson variable 'deb' for correct include directories.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>