meson: Clear up files() calls
[why]
files() does not need the files in an array.

In most (all?) of the use cases here we do not even need files(); but
keep it in because that would be needed if we have subdir meson.build
files; which would clean up all these hard coded relative path names.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>