Send messages when executing Steps and Sequences
[why]
Sequences should generally be executed in a separate thread,
but the main thread still needs to be informed about when
each step was started, stopped, and so on. This information
can be sent from the executing to the main thread via a
thread-safe message queue. This way, the Step and Sequence
classes do not need to be made thread-sage themselves.

[how]
This commit enhances the Message class and introduces a
send_message() function for conveniently enqueueing messages
in a queue. It enhances the execution functions for both
Step and Sequence by the ability to send messages when
steps/sequences are started or stopped.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>