Simplify execute_sequence_impl()
[why]
The previous implementation did not allow empty blocks
after WHILE, IF, and so on. It also needed to keep explicit
internal state for handling control flow.

[how]
Use a fully recursive implementation instead of the
previous semi-iterative one. A WHILE is now handled
by a while loop in the C++ code, an IF by an if(), a
TRY by a try, and so on.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>