Move execute_sequence() into Step class
[why]
execute_sequence() needs to modify each individual
step during the execution of the sequence. While
this is theoretically possible through the public
interface of Sequence, it would be rather complicated
and wasteful because Sequence would have to re-establish
its class invariants after each single modification.
As a member of the class, the function can directly
access the internal step list and modify steps as
necessary.

[how]
Move the code of the free function execute_sequence()
into a member function Sequence::execute().

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>