meson: Do not link GUL, leave that for the client
[why]
With possibly different versions of GUL around it can be hard if we use
one version of GUL while a user of taskomat has to use another version
of GUL.

[how]
We assume that the GUL version used to compile taskomat is 'about the
same' as the one the taskomat-client will use, and that the API of the
GUL objects we use in our interface stays unchanged.

As the affected types are only gul14::string_view and gul14::optional
that assumption is very realistic; these backports will probably never
change.

So:
- we do not link ourselves GUL anymore
- we put the linking instruction in the pkgconfig file
- we put the linking instruction in the taskomat_dep variable

[note]
We also use gul14::SlidingBuffer, but only in templates that are
evaluated in the client's compilation context, so we are safe there.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>