Add modify() function
[why]
To simplify the modification of a Step inside of a Sequence.

This function modifies one of the sequence steps in place. The modification is done
by a user-supplied function that receives a mutable reference to the Step indicated
via an iterator:

Sequence seq = get_sequence_from_somewhere();
auto it = seq.begin(); // get an iterator to the Step that should be modified

seq.modify(it, [](Step& step) { step.set_label("Modified step"); });

Background: A Sequence does not allow its steps to be modified directly from the
outside via references or iterators because it has to uphold certain class
invariants (like "all steps are always correctly indented"). A call to modify(),
however, reestablishes these invariants after the modification if necessary.

Proposed-by: Ulf Fini Jastrow <ulf.fini.jastrow@desy.de>
Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>