Step: Set last modification time on object creation
[why]
On Step creation the modification time is not set. Some steps like 'end'
will work with a empty created Step object of the correct type.

When we change the type of a Step the modification time is updated.
It is just that when creating a Step from a Type we do not use the
setter function (that updates the modification time).

It is a bit strange to update the time not always.

[note]
Also use 'one object one line' core guideline.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>