serialize: tests: Remove temp-dir only once
[why]
A lot of files are written to disk to check the serialization
deserialization process. To enable developers to have a look into the
generated files after the tests run, this directory is not removed after
the tests - but before the tests, to start with a fresh new empty
directory.

But that has been done on every test case, so that only the files from
the last test case can be inspected.

[how]
Instead of removing the temporary directory on each individual test case
it is now removed on test startup - once.

For this to work all stored sequences need to have unique names, one
sequence is therefore renamed.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>