serialize: tests: Centralize temp directory name
[why]
Avoid any mistakes when adding tests by putting the temporary directory
into a variable; that directory is removed on startup.

[how]
Use a const std::string to hold the directory.
It is a string to allow easy concatenation.
It is not constexpr because C++17 has not yet that strings.

[note]
Also sort #includes.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>