serialize: tests: Add test for filename escapes
[why]
When serializing a Sequence the directory it ends up in is derived from
the Sequence.label. But that might contain characters that are not
allowed in filenames (directory names).

Therefore these are escaped by a "$xx" sequence where xx is the
character's hex codepoint.

What is missing is a test, that this round-trips.

[how]
Just create a Sequence with a 'funny' name and restore it.

[note]
Created / pulles out helper to get directory contents.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>