Message: Stream Message::Type as text
[why]
To improve readability.

[how]
The strings that are used are a bit more tightly coupled to prevent
errors if more types will be added.

The operator<<(.., Type) is declared as friend, which it would not
need to be, but we need it defined before the operator<<(.., Message),
which is just impossible to do non-friend.
That is needed because the later uses the former.
Tricky. Have no better solution right now. So be it.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>