Executor: End task on destruction
[why]
When the Executor is destructed the thread it started is not (never)
joined and we end up with a zombie process.

[how]
When the Executor is gone we have no means to stop/join the execution
step anymore. We might have a CommChannel copy, but that does not do the
join(). So I think the execution thread NEEDS to be ended when the
Executor stops existing (i.e. in the destructor).

Having a destructor means we have no default move operators anymore,
so use rule of 5.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>