Sol: Tell Sol our Lua is exception safe
[why]
Sol does not really handle exceptions, as can be seen from
the linked Issue.

[how]
Tell Sol that our (custom built) Lua is exception aware.

See also
https://sol2.readthedocs.io/en/latest/exceptions.html#lua-and-luajit-c-exception-full-interoperability

[note]
Obviously (?) this worked before? At least the Doxygen comments seem to
bring the "[ABORT]" marked exceptions through. But that is impossible if
the exceptions are caught within Lua with `(...)` ... ??!

Fixes: #10

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>