time_types: Make TimePoint dump thread safe
[why]
std::gmtime() is not thread safe as it uses one static storage location
for all concurrent calls to the function.

[how]
Use the C function gmtime_r() instead, that uses a user-provided output
buffer.
On Windows the function is called gmtime_s() and takes the arguments in
swapped order. The return value types also differ.

[note]
We have no Windows CI stuff set up at the moment, so this is an untested
stub for Windows.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>