time_types: Add timegm()
[why]
For deserialization we need a conversion function from struct tm to
time_t, that works in UTC. mktime() is the reverse of localtime().
There is no library function that performs that task and only
nonstandard ways to enable mktime() to perform in UTC.

[how]
Create a new function as reverse to gmtime(): timegm(). That function is
available on some platforms from their standard library; but not on all
platforms.

The function does neither check nor correct the data stored in the struct tm.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>