Revert "Sol: Tell Sol our Lua is exception safe"
This reverts commit 6026126e30d5340bc45af3991346a5e011d373b8.

The change introduces other errors, for example when a doocs EqCall
throws it is not chaught. Probably the setting means 'we promise we will
never throw'?

This throwing stuff inside Lua/Sol is strange.

And I can not reproduce my own issue #10, which was the reason for the
now reverted commit.