time_types: Fix test for systems with 32b time_t
[why]
With 32b time_t taking twice the amont of std::time(nullptr) can easily
be too far in the future and be wrapped back to something we do not
want. Go to two times now into the future but not more than the max.

[how]
If the value in time_t wraps on multiplication we can not detect it with
the simple `std::time(nullptr) > std::numeric_limits<std::time_t>::max()`
Instead we must compare this before the multiplication.