Sequence: Jump over disabled Steps/Step-blocks
[how]
Only 'block initializing steps' can be disabled. These are ACTION, IF,
WHILE, and TRY steps. Other Steps, that (for a syntax error free
Sequence) need one of these initializing block before them, can not be
diabled on their own.

So one can only disable complete IF-ELSE blocks, or TRY-CATCH blocks. It
would be meaningless to disable the IF block but not the ELSE block.

So we ignore all Steps in/under the current block (if it has any), and
then we ignore all further blocks that are dependend (not initializing)
blocks that follow on the same indentation level.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>