Sequence: Fix disabling of more Steps
[why]
When one Block-including-its-continuation is disabled the next Block
can not be ignored. This is a bug, I saw it right when reading the MR on
Gitlab :-(

So the code needs (assumes) at least one not-disabled Step between
disabled Blocks.

[how]
After finding the end of a Block-including-its-continuation the
next-to-be-executed Step is checked for disabled-ness before it is
executed.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>