deserialize: Optimize code
[why]
I'm not really sure why, but when I added the parsing for 'disabled' the
comment was to use string_views.

The problem is, that the original design itself is already broken in
this regard. To make everything worse it searches for a colon two times
where one time is enough.

It also searches for the colon and a potential keyword when the script
part of the save file has already been reached.

So this is more or less a rewrite that solves
* do not parse the savefile after the script has begun
* split keyword and value only once
* use string_views where possible

[how]
In the beginning we have a line. Here the first colon is searched for,
and the left part (after expecting the line to start with double dash)
is assumed to be the keyword.
We know now where the split is, and modify the line-rest that it just
contains the value. In this way we avoid the double parsing of the line
for a colon.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>