Sequence: Propagate 'disabled' into sub steps
[why]
It is easier to visualize if not only the control structure step (e.g.
Step::type_if) is set to disabled and silently disabling enything up to
the next Step::type_end, but to actually set the disabled properties in
all those steps.

[how]
When we check the indentation level we walk the complete structure and
can there also check validity and autocorrect of all disabled flags.

[note]
Re-enabling a control structure just enables it and not any of the
contained substeps.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>