Sequence: Reenable complete control structures in one go
[why]
The previous commit added the invariant that if a control structure is
disabled all its contained steps are also disabled like so:

(Ticked steps are disabled)

[ ] ACTION
[ ] IF ...
[ ]     ACTION
[ ] END
[ ] ACTION

If the 'IF' is disabled all what belongs to it is automatically also
disabled. Left side is the 'user action', right side is after invariant
enforcement.

[ ] ACTION               [ ] ACTION
[+] IF ...               [X] IF ...
[ ]     ACTION           [X]     ACTION
[ ] END                  [X] END
[ ] ACTION               [ ] ACTION

But it just checks the invariant and does not know which step has been
changed. So if the IF step is re-enabled, we have this:

[ ] ACTION               [ ] ACTION
[o] IF ...               [ ] IF ...
[X]     ACTION           [X]     ACTION
[X] END                  [ ] END
[ ] ACTION               [ ] ACTION

(The IF is checked: Is enabled: ok. The ACTION is checked: Is disabled:
ok. The END is checked: Is disabled still: Not ok, correcting)

But the expected behavior is not just keeping the invariant, but also
change subsequent steps in some way if one particular step has been
changed: All subordinary steps should FOLLOW the control step like so:

[ ] ACTION               [ ] ACTION
[o] IF ...               [ ] IF ...
[X]     ACTION           [ ]     ACTION
[X] END                  [ ] END
[ ] ACTION               [ ] ACTION

(The IF is re-enabled, so all substeps are also re-enabled)

[how]
This is a bit more involved than the previous commit. This needs the
information WHICH concrete step has been changed and can not work on the
Sequence as a whole without that knowledge.

Although it would have been possible to traverse the changed control
structure and all its subsequent steps and set them all to 'enabled',
this would lead to duplication of the traversal code.

Instead the normal 'enforce the invariant' function indent() gets a new
parameter that, if set, conveys the information about which step should
be the 'master' for its subsequent steps.

The modify() function works on iterators, and to keep it simple update()
switches from a loop of Steps to a loop of Step iterators.

If indent() is called with iterator parameter it's behavior is changed
only after the step the iterator points at. It will keep the current
disabled-flag until the end of the current control structure. After that
has been traversed the iterator is set to end() and thus disabled (i.e.
normal behavior of indent() for the remaining steps).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>