Replace tests with race conditions by simple is_running() test
[why]
The two tests removed by this commit created a sequence,
started it in a worker thread, and at the same time
attempted to modify it in the main thread. Since there
was no synchronization whatsoever, this was a race condition
and the test behavior was therefore undefined.

[how]
Replace the two tests by a simple test for the
initialization and condition after execute() of
is_running().

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>