Use string_view as argument type in send_message()
[why]
The previous pass-string-by-value-then-move-out-of-the-argument
scheme was, arguably, a bit complicated. More importantly,
except for a single instance in this library where the function
was used with an inline cat(), it did not bring any benefit.
One might argue that string_view is the better argument type for
this usage. Or not. :)

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>