Embed Unicode symbols in abort marker
[why]
The abort marker in error messages used to be
very easy to use accidentally by user error
messages ("[ABORT]"). Something less common
seems to be in order.

[how]
Use the text "ABORT" surrounded by two UTF-8
stop signs as an abort marker. While we're at
it, surround the error message with the marker
instead of just prefixing it. This will later
allow a precise retrieval of only the original
error message.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>