Produce nice error messages on sequence abort
[why]
So far, the messages produced when aborting sequences
were hardly readable:

2022-08-30T16:38:52 script output [0]: Sequence stopped with error: Error while executing script of step 3: sol: runtime error: caught (...) exception stack traceback: [C]: in function 'base.sleep' [string "print('wait 2s and poll button again')..."]:2: in main chunk

[how]
Add special handling for the case that the sequence
was aborted by the user (as opposed to "stopped with
an error"). Carefully avoid adding too much text in
the various error handling layers and make the remaining
text more concise. Add a test to ensure that the user
sees only the message:

"Sequence aborted: Stop on user request"

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>