Change type of abort_marker to string_view
[why]
For std::string, a constructor has to be called in the
library initialization phase. It may allocate a few
bytes and it copies the string.

For const char*, there is no constructor and no
additional memory consumption, but on every usage
strlen() has to be called because the length of the
string is not stored anywhere.

For string_view, there is a constructor call which runs
strlen() once, but there is no allocation, no copy, and
the object only consumes an additional 16 bytes
(2 pointers). For later usage, the length is already
known.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>