Add explicit sequence termination with a custom Lua function.
[why]
There is a need to explicitly terminate a sequence with a custom Lua
function (here called 'terminate_sequence()') in a Step. It exit a
sequence gracefully and returns silently from the Sequence::execute()
member function.

[how]
I implemted a flag in Step class when the custom Lua function for
termination is called. It will throw an Error exception with a
special string instance. The Sequence execute member function
will handle it in a special way and exit the sequence silently.