Simplify explicitly terminating a sequence.
After a discussion with Lars there is no need to have a special marker
to terminate the sequence as we can also use the abort marker and
function. This simplifies the code a lot because there is neither a
need to have a special flag for termination nor a special termination
marker.