Add tests for user exceptions
[why]
A recent commit broke the exception handling for user code (again).

[how]
Add a test where the user creates a Context with a new Lua function that
throws an exception. Assert that the exception is caught and somehow
meaningful processed (i.e. visible for the script writer).

This is related to Issue #4:
  https://github.com/taskolib/taskolib/issues/4

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>