Step: Prevent labels that we can not serialize
[why]
The serialisation process is designed to accept any number of
whitespaces after the colon. This will suppress any whitespaces that
were part of the label to be restored correctly.

[how]
Just forbid labels that can not be restored. Although we could instead
install several other attempts to allow leading or trailing whitespaces
in the label the question is: Is the freedom to allow such labels worth
the more complicated code?
For this commit the answer is: no, and so whitespaced labels are
coerced (i.e. trimmed).

Fixes:
https://mcs-gitlab.desy.de/doocs/doocs-high-level-support/libtaskomat/-/issues/19

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>