Add ErrorAtIndex exception class; rename files
[why]
We need an exception class to transport the index
of the step in which an error happened.

[how]
Add an ErrorAtIndex class that is derived from Error
to the header file. As this introduces a second class
into the header, also rename Error.h to exceptions.h.

Also: Introduce a type for step indices in exceptions.h.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>