Use optional<StepIndex> in messages and out functions
[why]
Since we can now transport errors with or without
associated step index information, this must also
be reflected in the messages passed and in the
output functions.

[how]
Use optional<StepIndex> wherever we used a simple
StepIndex before except in places where we are sure
that a step index is available.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>