Executor: Make post-run context available
[why]
When we execute a Sequence directly we can access the Context
afterwards, for example to extract calculated variable values.

This is not possible for asynchronously executed Sequences because their
Context is never synced back.

[how]
When a Sequence is finished use the future to retrieve the processed
Context and store it in the Executor. The user can retrieve the updated
variables with the get_context_variables() call.

Mentioned in: #24

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>