Rename to taskolib
Replacing the dev package, but old and new library packages can be
installed as customary with Doocs.

[note]
The include path changed from <taskomat/*> to <taskolib/*>.
The pkgconfig dependency is now called `taskolib`.

[note]
With commit
  c333841  Rename Taskomat class to SequenceManager.

the header Taskomat.h has been renamed to SequenceManager.h but this
change has been forgotten in taskomat.h. This is fixed en passant.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>