Step: Resync variables from state to context also on errors
[why]
If a Step has an error (syntax or runtime), or aborts changed
variables from the context are not synced back to the context.

This makes it impossible to use the variables that are handed over from
Step to Step in a Sequence to keep information stored into them if the
Step that stored the information errors out.

[how]
Copy the context variables after the script returned from state to
context always. Regardless if the script errored out or not.

Fixes: #24

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>