Use execute_lua_script_safely() inside Step::execute()
[why]
To intercept exceptions coming out from Sol and to
handle them correctly.

Note: As a side effect of using execute_lua_script_safely(),
this commit changes the handling of return values from the
scripts in an unintended way (from Lua-style boolean decay
to C-style boolean decay). This is addressed in the next
commit.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>