Require bool return value for IF/ELSEIF/WHILE
This commit divides the step types into two categories:

For IF, ELSEIF, WHILE, requires_bool_return_value() is true
and not returning a value from the script is considered an
error, as is returning a value of non-bool type.

For the other step types (like ACTION),
requires_bool_return_value() == false and returning any
value but nil from the script is considered an error.

Unit tests are extended accordingly; some minor modifications
to existing tests are also required.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>