Step: Add bool and nil context variable types
[why]
These two basic Lua types can not be forwarded from one step to the
next.

[how]
Add the bool type just like the others.

Nil in Lua means 'does not exist', so `x = nil` means that the variable
x becomes non-existing afterwards. That means that we need to remove the
variable also from the context in that case. We do not need a 'nil
type'; nil is just the absense of anything there.

The Lua types 'userdata', 'function', and 'table' can still not be
forwarded from one Step to the next.

Fixes: #32

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>