meson: Remove include kludge
[why]
The previous commit
  7def9b5  meson: Fix configuring for meson < 0.56

introduced a not very robust code passage to manipulate the strings of
the included files. This is because Meson does not have the needed
string and/or array manipulation features in the required version.

[how]
Externalize the Meson include setup into its own file. This allows to
use the 'correct' pathnames right from the start and no string
operations are needed at all.

The root meson.build looses more of its centralism, which
 * can make the build setup less clear (less centralized)
 * reduce possibility of merge conflicts
 * improve singe-purposeness of the individual meson files

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>