meson: Add wrap for GUL
[why]
We depend on GUL. If someone wants to use taskolib and has not yet
anything to do with GUL, that means building and installing yet another
lib beforehand etc

[how]
Allow to build GUL as subproject.

[note]
The dependency() extra-arguments can be removed when we update to
Meson 0.55.

[note]
Maybe we should push GUL into wrapdb?

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>