Simplify implementation of terminate_sequence()
[why]
The previous implementation captured the Sol state by
reference at the moment of registering the function with
Lua, and then used .lua_state() to access the Lua
state.

[how]
It is slightly more idiomatic and less error-prone to use
Sol2's sol::this_state type as a parameter and to make
use of the implicit conversion from Sol states to Lua
states.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>