Context: Step: Fix direct context variable asignment with char*
[why]
On some (older) compilers std::variant<> in 'broken' in this way:
  If the variant can hold std::string and bool, and it is passed a char*
  this will bind as bool variant and not the expected string variant.
  The reason is that a char* is first a pointer. A pointer is an int
  number. An integer converts to bool. So the user must mean bool?
  See also [1].

Tested on Ubuntu 20.04 this is the case (GCC 9.4, clang 10.0).
Tested on Ubuntu 22.10 this is fixed (GCC 12.2, clang 15.0)

This only happens if a user does a direct assignment into the context's
variable member. Steps export the variable always as std::string.

This only appeared after `bool` has been added to the possible variable
types with commit
  cacd840  Step: Add bool and nil context variable types

Before that, because there is no bool, the compiler created a
std::string from the char*.

[how]
If we want to allow direct char* assignment to the context variables
(and we should, because we can not effectively forbid it), the only
option to disembugiate is to allow char* to be stored in the variables.

This has the drawback, that the data at which the char* points needs to
be valid until we finally create a std::string from it. That only
happens when the context variable is imported into Lua while executing a
Step. When the variable leaves the Step it will stay a std::string.

The solution is almost transparent for the user, the only thing to keep
in mind is lifetimes. The pointer is not immediately converted (i.e.
copied) into a std::string but only later, far later sometimes.
But because that is an expert option in an expert option (the
documentation says you shall assign std::strings and not char*), this
might be acceptable.

[1] https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2018/p0608r3.html

Reported-by: Lars Froehlich <lars.froehlich@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>