Simplify Github build & test action
[why]
The ci.yml is an unnecessary complication for dispatching the sub action
to build & test. It is not needed.

[how]
We have now only one file that gathers all necessary steps to build and
test on Ubuntu 20 when pushed to main branch.

[issue]
Still need to solve the problem with duplicated ignore files.