Report weird exceptions as "Unknown exception"
[why]
If Sol3 intercepts a Lua/C++ exception and there is no
error message string on the Lua stack, it reports a
weird error message that we previously converted to
"Unknown C++ exception". That is not entirely true
because Lua code can generate the same thing.

[how]
Report "Unknown exception" if it is unclear whether
the error came from Lua or C++ code. Keep reporting
"Unknown C++ exception" only if we caught an actual C++
exception via catch(...).

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>