Make Step::execute() more exception safe
[why]
During the preparation of the Step environment, other
things can go wrong than just the user script itself.
For example, the LUA init function might throw. In this
case, we need to make sure that both a "step started"
and a "step stopped with error" message are sent,
otherwise the step remains marked as "running" on the
Executor side.

[how]
Factor out an inner function Step::execute_impl() that
may throw and call it from the outer function
Step::execute() that takes care of sending messages
and catching the exceptions.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>