Test only strictly guaranteed conditions
[why]
The "run sequence asynchronously with throw" test
for the Executor class tested for Executor::is_busy()
directly after an Executor::run_asynchronously() call.
While it is very likely to return true, this is not
guaranteed and actually fails on some Github test
systems.

[how]
Test only for conditions that are independent of
timing.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>