Introduce aliases for Lua types
[why]
When we want to interface with the Lua data types that are held in the
Lua state we need to use specific C++ types. These types (for example
`long long` have been used directly and often it is not so clear why
some type is used.

[how]
Use alias types that clearly communicate to the reader of the case that
some value for interfacing with Lua is to be used/created/stored.

When using these also consistently to create assignments the whole
problem of variant's char*-to-bool quirk goes away. We could even
introduce not alias types but real types instead to be even more type
save.

[note]
Also unify the test code a bit.

Also change how the values in the Context.variables are compared
(sometimes). We compare the values, not the variants.

Also add missing LuaBool export from Script test.

Also add missing funtion export from Script test.

Fixes: #47

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>