Introduce aliases for variable storage types
[why]
Sometimes it is useful to know why a concrete type has been choosen, and
which other type is connected to it.

We have a lot of types that need to be a specific one. To show the
interdependence we introduced the Lua type names in the previous
commit. But that is only half of the truth. The types in the Context
variables storage (VariableTable) can in general be of different type,
for example when other scripting languages are added etc. They must be
used to std::get the values out of the storage.

[how]
This cleans up the distinction between 'is a given because of Lua' and
'we choose this type for storage' (which is the same type as Lua's
types, but the reason is just lazyness and avoiding casting.

For this we introduce another set of 4 types and use them consistently
whenever we use the variable storage in Context. (The Lua types are used
when accessing the variable storage of Lua.)

[note]
Also remove unneeded case in Step::copy_used_variables_from_context_to_lua()
that expected a type in the VariableTable that can never be there (nil).

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>