Add Timeout class
[why]
The special value for "infinite" timeouts makes the
handling of std::chrono durations very error-prone
for Taskolib users. It may be better to introduce
a custom class for timeouts that automatically takes
care of conversions to floating-point values and
similar special cases.

[how]
Introduce a class Timeout that can be constructed
from std::chrono durations or from floating-point
values (interpreted as number of seconds). The class
supports implicit conversion to floating-point and
explicit conversion to std::chrono durations (because
the latter has no obvious way of representing
infinity). The usual comparison operators are also
supported.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>