Executor: Make is_busy() private
[why]
is_busy() just checks for the status of the parallel thread
and joins it if necessary, but it does not process any
messages from the queue. However, the entire working of the
executor depends on regular message processing, so we may
not tempt users to write code like this:

while (executor.busy())
    sleep(100ms);

At the end of this loop, the sequence will have executed,
but not a single message has been processed, and the local
copy of the sequence does not reflect any of the execution
state. In the worst case, the message queue is full, and
execution blocks indefinitely.

[how]
Make is_busy() private and treat it as an implementation
detail. Users are now forced to regularly call update().

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>