Sequence: Return optional Error objects with get_error()
[why]
If a sequence is stopped prematurely, users should not just
be able to retrieve the associated error message, but the
entire exception information including the (optional) step
index of the problem.

[how]
Replace get_error_message() by get_error(), which returns an
optional<Error>. It is nullopt to indicate that the sequence
finished successfuly, and it contains an Error object if it
stopped prematurely. Adapt all usages accordingly.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>