Consistently use OptionalStepIndex instead of StepIndex
[why]
In the 1-argument version of Step::execute(), we still
used a fixed step index of 0 to indicate "no step information".
This is in contradiction with the use of an OptionalStepIndex
throughout the higher levels of the code.

[how]
Unify the two Step::execute() overloads into one which
takes an OptionalStepIndex and consistently use
OptionalStepIndex in all lower-level functions used
by it.

Signed-off-by: Lars Froehlich <lars.froehlich@desy.de>